requirements.txt
streamlit
ai71
python-dotenv